refactor: unify user data access interfaces
