--- src/main/java/use_case/logout/LogoutUserDataAccessInterface.java
@@ -1,21 +1,0 @@
-package use_case.logout;
-
-/**
- * DAO for the Logout Use Case.
- */
-public interface LogoutUserDataAccessInterface {
-
-    /**
-     * Returns the username of the curren user of the application.
-     * 
-     * @return the username of the current user
-     */
-    String getCurrentUsername();
-
-    /**
-     * Sets the username indicating who is the current user of the application.
-     * 
-     * @param username the new current username
-     */
-    void setCurrentUsername(String username);
-}
